Major re-work of Virtual Serial Channels
Some bugs were introduced in virtual serial channels around a year ago.
Also the original code was "hackity" and hard to maintain.  This change is
a complete re-write of DWVModem which uses a state machine to keep track of
what the virtual modem is doing.  This significantly reduced the code
complexity.  Completely tested and working with NitrOS-9.