(issue #7) Do some clean up when server sockets close
Clients connecting to pyDriveWire over a socket can terminate suddenly.  If
the clients opened Virtual Serial channels those can remain open and may
cause problems when the client reconnects.  The fix adds a callback
mechanism to DWServerSocket and DWSimpleSocket when the connection closes.
The callback is set to the Server's cmdInit method.  This resets all the
open disk files and closes all of the Virtual Serial channels.