Small Fixes: SSH Enabled Message; Allow Experimental in Config; Allow Spaces in Config
The config file now properly processes 'option experimental' lines.  These can
be in addition to supplying them on the command line as well.

    option experimental printing
    option experimental ssh

These are equivalent to supplying

    -x printing
    -x ssh

on the command line

The config file format is adjusted as follows to allow for spaces:

    option <optionName> <value>

Value is now treated as the remainder of the line, including spaces.