Add a mode to support DriveWire on the DosPlus Extender
The DosPlus Extender for Dragon computers has DriveWire support.
The extender seems to be implemented for DriveWire servers which
do not provide proper support for VDK files.  When the Dragon
makes a request for a particular LSN, the ReadEx and Write
commands sent by the extender are actually LSN+1. This seems
to be because other DriveWire servers do not have full support
for the VDK disk image format.  pyDriveWire will properly
detect a VDK image file and never send the VDK header to the
client.  When pyDriveWire's dosplus mode is enabled the
reverse LSN = LSN-1 is applied to correct this.

The dosplus flag is on a per-disk image basis.  For convenience
there is also a server-wide settting.  Commands added:

Insert a disk with the dosplus workaround enabled:

    dw disk insert <drive> <path> --dosplus

Query whether dosplus is enabled for a drive:

    dw disk dosplus <drive>
    dw disk info <drive>

Change the dosplus flag for a given drive:

    dw disk dosplus <drive> <setting>

        Setting can be: 0, f[alse], n[o], on
                        1, t[rue], y[es], off

Query the server global dosplus default:

    dw server dosplus

Change the dosplus flag for a given drive:

    dw server dosplus <setting>

        Setting can be: 0, f[alse], n[o], on
                        1, t[rue], y[es], off

Enable the server global dosplus default from the command line:

    ./pyDriveWire --dosplus ...